+++ docs/misc/changelog.rst
@@ -19,6 +19,7 @@ New Features:
 - Add support for Callback (cf https://github.com/hill-a/stable-baselines/pull/644)
 - Add methods for saving and loading replay buffer
 - Add `extend()` method to the buffers
+- Add `get_vec_normalize_env()` to `BaseRLModel` to retrieve `VecNormalize` wrapper when it exists
 
 Bug Fixes:
 ^^^^^^^^^^